requirements.txt
python-docx>=0.8.11
markdown>=3.4.0
beautifulsoup4>=4.11.0
dify-plugin